==> 2) Finding Articles with Program Streams.ipynb <==
# Exploring the Press Data
## 2) Finding Articles with Program Streams

import numpy as np
import pandas as pd

import pprint
import json
import bson

Define the functions to search for words in the list of articles and create a dictionary to store matches

# Determines whether the word occurs within the given list of strings
# returns True if so, and False if not
def is_word_in_string_list(list_of_strings, word):
    for line in list_of_strings:
        if line.find(word) != -1:
            return True
    return False

# Creates a dictionary of the list of articles that contain each word in the given list of words
# NOTE: this function assumes the incoming article objects have the same format as the raw articles stored 
# on MongoDB. That is, the full text is stored as a list of strings for each article object
# and can be accessed from the article object using "article_name['details']['fulltext']"
# If duplicate words are contained within the list of words, the function issues a warning,
# detailing which word occurs twice, but still continues processing
# input: list of search words, list of article objects
# output: dictionary of the list of articles containing each search word
def create_dictionary_of_articles(list_of_words, list_of_articles):
    
    article_dictionary = dict()
    
    # search for each word in the list of articles
    for word in list_of_words:
        
        # sanity check to make sure the list of words doesn't contain duplicates
        if word in article_dictionary:
            print("Warning: found a duplicate!", word)
            continue
        
        article_dictionary[word] = []
        
        # search through the articles for the word, and append the article to the dictionary if the word is found
        for article in list_of_articles:
            
            text = article['details']['fulltext']
            
            if is_word_in_string_list(text, word):
                article_dictionary[word].append(article)
        
        print("Found", len(article_dictionary[word]), "articles for", word)
        
    return article_dictionary

Get the list of program streams from the Programs excel spreadsheet

# read the Programs sheet
programs_list = pd.read_excel("./excel_sheets/Departments and Programs.xlsx", sheet_name = 'Programs')['Programs'].to_numpy()
programs_list = np.sort(programs_list)

print("Number of programs:", len(programs_list))

Import the cleaned articles from the cleaned_articles.json file in the "data" directory, and convert the JSON objects into BSON objects for possible integration with MongoDB

with open('./data/cleaned_articles.json') as json_file:
    data = json.load(json_file)
    articles_json = data['cleaned_articles']

articles = []

# import function needed to convert JSON into BSON
from bson.json_util import loads

# convert the article from JSON into BSON
for article in articles_json:
    articles.append(loads(article))

Search for the program names in the list of articles, and store the results in a dictionary

results = create_dictionary_of_articles(programs_list, articles)

Exporting the results to a JSON file called *cleaned_articles_dictionary.json* in the **data** directory

# import function needed to convert BSON into JSON
from bson.json_util import dumps

# create a dictionary using JSON instead of BSON
results_json = dict()
for program in results:
    results_json[program] = []
    for article in results[program]:
        results_json[program].append(dumps(article))

data = dict()
data['cleaned_articles_dictionary'] = results_json

# save to the file (note, this re-writes the entire file)
with open('./data/cleaned_articles_dictionary.json', 'w') as outfile:
    json.dump(data, outfile)
